Clean one more

+++ js/common/lib/tensor.ts
@@ -4,7 +4,6 @@
 import {TensorFactory} from './tensor-factory.js';
 import {Tensor as TensorImpl} from './tensor-impl.js';
 import {TypedTensorUtils} from './tensor-utils.js';
-import {Float16Array} from '@petamoriken/float16';
 
 /* eslint-disable @typescript-eslint/no-redeclare */
 
@@ -64,6 +63,7 @@ interface TypedTensorBase<T extends Tensor.Type> {
   dispose(): void;
 }
 
+// type Float16ArrayType = InstanceType<typeof Float16Array>;
 export declare namespace Tensor {
   interface DataTypeMap {
     float32: Float32Array;
@@ -75,7 +75,7 @@ export declare namespace Tensor {
     int64: BigInt64Array;
     string: string[];
     bool: Uint8Array;
-    float16: Float16Array;  // Keep using Uint16Array until we have a concrete solution for float 16.
+    float16: Uint16Array;  // Keep using Uint16Array until we have a concrete solution for float 16.
     float64: Float64Array;
     uint32: Uint32Array;
     uint64: BigUint64Array;
